requirements.txt
Flask
numpy
pandas
scikit-learn
nltk
spacy
contractions
matplotlib
seaborn
wordcloud
keras-tuner
tqdm
tensorflow
gensim

